# src/client/generated/w3s-openapi.yaml
openapi: 3.0.3
info:
  title: W3S
  description: All W3S APIs
  version: '1.0'
servers:
  - url: https://api.circle.com/v1/w3s
security:
  - BearerAuth: []
...
paths:
...
  /contracts/import:
    post:
      tags:
        - Deploy / Import
      summary: Import a contract
      description: Import contract
      operationId: importContact
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportContractRequest'
        description: Import contract request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ContractResponse'
          description: Idempotent request. Contract already imported
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        '201':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ContractResponse'
          description: Contract imported
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
  /contracts/deploy:
    post:
      tags:
        - Deploy / Import
      summary: Deploy a contract
      description: Deploy a smart contract
      operationId: deployContract
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractDeploymentRequest'
            examples:
              Wallet ID:
                value:
                  idempotencyKey: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
                  name: First Contract
                  description: My first hello world contract
                  walletId: 004735f6-d9fc-44f8-933c-672cdf3d240d
                  abiJson: "[\n\t{\n\t\t\"inputs\": [],\n\t\t\"stateMutability\": \"nonpayable\",\n\t\t\"type\": \"constructor\"\n\t},\n\t..."
                  bytecode: 0x60806040523480156200001157600080fd5b50604051806040...
                  constructorParameters:
                    - TICK
                    - 10000
                  feeLevel: MEDIUM
                  entitySecretCiphertext: 0NtD3d3+nmgb4GqYQXzAjKF8h5Zq6sHM2k/7UQz5OIXWnU7H8O9XvVYXqDnw1wEmY...
        description: Deploy contract request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ContractDeploymentResponse'
          description: Idempotent request. Contract deployment request already submitted
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        '201':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ContractDeploymentResponse'
          description: 'New contract deployed '
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
  /contracts/deploy/estimateFee:
    post:
      tags:
        - Deploy / Import
      summary: Estimate a contract deployment
      description: Estimate the transaction fee for deploying a smart contract
      operationId: estimateContractDeploy
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractDeploymentEstimateFeeRequest'
            examples:
              Wallet ID:
                value:
                  walletId: 004735f6-d9fc-44f8-933c-672cdf3d240d
                  bytecode: 0x60806040523480156200001157600080fd5b50604051806040...
                  constructorSignature: constructor(string, uint256)
                  constructorParameters:
                    - TICK
                    - 10000
              Source Address:
                value:
                  sourceAddress: '0x1bf9ad0cc2ad298c69a2995aa806ee832788218c'
                  blockchain: MATIC-MUMBAI
                  bytecode: 0x60806040523480156200001157600080fd5b50604051806040...
                  constructorSignature: constructor(string ticker, uint256 totalSupply)
                  constructorParameters:
                    - TICK
                    - 10000
        description: Estimate a transaction fee for deploying a smart contract
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ContractDeploymentEstimateFeeResponse'
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
...
components:
  schemas:
    ContractDeploymentEstimateFeeRequest:
      type: object
      required:
        - bytecode
        - constructorParameters
      properties:
        blockchain:
          $ref: '#/components/schemas/commonBlockchain'
        bytecode:
          $ref: '#/components/schemas/Bytecode'
        constructorSignature:
          $ref: '#/components/schemas/ConstructorSignature'
        constructorParameters:
          $ref: '#/components/schemas/ConstructorParameters'
        sourceAddress:
          $ref: '#/components/schemas/SourceAddress'
        walletId:
          $ref: '#/components/schemas/WalletId'
    ContractDeploymentEstimateFeeResponse:
      type: object
      properties:
        high:
          $ref: '#/components/schemas/modelTransactionFee'
        low:
          $ref: '#/components/schemas/modelTransactionFee'
        medium:
          $ref: '#/components/schemas/modelTransactionFee'
    ContractDeploymentRequest:
      type: object
      required:
        - abiJson
        - bytecode
        - entitySecretCiphertext
        - idempotencyKey
        - name
        - constructorParameters
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
        name:
          $ref: '#/components/schemas/ContractName'
        description:
          $ref: '#/components/schemas/ContractDescription'
        walletId:
          $ref: '#/components/schemas/WalletId'
        abiJson:
          $ref: '#/components/schemas/AbiJson'
        bytecode:
          $ref: '#/components/schemas/Bytecode'
        entitySecretCiphertext:
          $ref: '#/components/schemas/EntitySecretCiphertext'
        blockchain:
          $ref: '#/components/schemas/commonBlockchain'
        constructorParameters:
          $ref: '#/components/schemas/ConstructorParameters'
        feeLevel:
          $ref: '#/components/schemas/modelFeeLevel'
        gasLimit:
          type: string
          description: GasLimit is the limit on how much gas can be used in the transaction.
        gasPrice:
          type: string
          description: GasPrice is the price per unit of gas for the transaction.
        maxFee:
          type: string
          description: MaxFee is the total amount willing to be paid per unit of gas for your transaction.
        priorityFee:
          type: string
          description: |-
            PriorityFee is an additional fee per unit of gas to incentivise validators to complete your transaction.
            faster.
        refId:
          $ref: '#/components/schemas/RefId'
    ContractDeploymentResponse:
      type: object
      properties:
        contractId:
          type: string
          description: Unique identifier of the created smart contract.
          example: 676f83a8-81a1-4dd5-b738-e2509b8f5460
        transactionId:
          type: string
          description: Unique identifier of the pending deployment transaction.
          example: 36790743-b78d-4061-9558-1af9a9c837bd
    Contract:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/schemas_Id'
        deployerWalletID:
          type: string
          format: uuid
          description: The `id` of the Programmable Wallet that deployed this contract.
          example: f39e3dbd-84af-4d3d-b5ac-98fbe047ce6a
        deploymentTransactionId:
          type: string
          format: uuid
          description: The id of the deployment transaction for this contract.
          example: 00ca46f3-c31b-4e3a-92a9-068d4ff26a63
        implementationContractId:
          type: string
          format: uuid
          description: If this contract is a proxy, the contract ID of the implementation.
          example: f3d81192-1f5d-4d17-ae85-5801a15ea273
        txHash:
          type: string
          description: The string value of the transaction hash.
          example: '0x4a25cc5e661d8504b59c5f38ba93f010e8518966f00e2ceda7955c4b8621357d'
        abiJson:
          $ref: '#/components/schemas/AbiJson'
        archived:
          $ref: '#/components/schemas/Archived'
        blockchain:
          $ref: '#/components/schemas/Blockchain'
        bytecode:
          $ref: '#/components/schemas/Bytecode'
        contractAddress:
          type: string
          description: The on-chain address of this contract.
          example: '0x1e124d7384cd34448ea5907bd0052a79355ab5eb'
        contractInputType:
          $ref: '#/components/schemas/modelContractInputType'
        deployerAddress:
          type: string
          description: The address that created this contract, if deployed.
          example: '0x1bf9ad0cc2ad298c69a2995aa806ee832788218c'
        name:
          type: string
          description: The contract's name.
          example: First Contract
        description:
          $ref: '#/components/schemas/ContractDescription'
        status:
          $ref: '#/components/schemas/modelStatus'
        verificationStatus:
          $ref: '#/components/schemas/modelVerificationStatus'
        metadataLink:
          type: string
          description: The ipfs metadata link of the contract.
          example: https://ipfs.io/ipfs/Qme7ss3ARVgxv6rXqVPiikMJ8u2NLgmgszg13pYrDKEoiu
        updateDate:
          $ref: '#/components/schemas/UpdateDate'
        createDate:
          $ref: '#/components/schemas/schemas_CreateDate'
        sourceCode:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/modelSolFile'
        functions:
          type: array
          description: Functions supported by this contract. Parsed from abi_json.
          items:
            $ref: '#/components/schemas/modelFunction'
        events:
          type: array
          description: Events this contract can emit. Parsed from abi_json.
          items:
            $ref: '#/components/schemas/modelEvent'
    ContractResponse:
      type: object
      properties:
        contract:
          $ref: '#/components/schemas/Contract'
...
    ImportContractRequest:
      type: object
      required:
        - idempotencyKey
        - name
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
        name:
          $ref: '#/components/schemas/ContractName'
        description:
          $ref: '#/components/schemas/ContractDescription'
        address:
          type: string
          description: Address is the address of the contract to be imported.
          example: '0x1e124d7384cd34448ea5907bd0052a79355ab5eb'
        blockchain:
          $ref: '#/components/schemas/commonBlockchain'
...
    commonBlockchain:
      type: string
      description: Blockchain is the blockchain to deploy onto.
      enum:
        - ETH
        - ETH-GOERLI
        - AVAX
        - AVAX-FUJI
        - MATIC
        - MATIC-MUMBAI
      example: MATIC-MUMBAI
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
        walletId:
          $ref: '#/components/schemas/WalletId'
        abiFunctionSignature:
          $ref: '#/components/schemas/AbiFunctionSignature'
        entitySecretCiphertext:
          $ref: '#/components/schemas/EntitySecretCiphertext'
        abiParameters:
          $ref: '#/components/schemas/AbiParameters'
        blockchain:
          $ref: '#/components/schemas/commonBlockchain'
        feeLevel:
          $ref: '#/components/schemas/modelFeeLevel'
        gasLimit:
          type: string
          description: GasLimit is the limit on how much gas can be used in the transaction.
        gasPrice:
          type: string
          description: GasPrice is the price of gas for the transaction.
        maxFee:
          type: string
          description: MaxFee is the total amount willing to be paid for your transaction.
        priorityFee:
          type: string
          description: PriorityFee is an additional fee to have your transaction completed faster.
        refId:
          $ref: '#/components/schemas/RefId'
    modelContractInputType:
      type: string
      enum:
        - IMPORT
        - BYTECODE
        - TEMPLATE
    modelEvent:
      type: object
      properties:
        name:
          type: string
          example: Approval
        type:
          type: string
          example: event
        anonymous:
          type: boolean
          example: false
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/modelParamType'
    modelFeeLevel:
      type: string
      enum:
        - LOW
        - MEDIUM
        - HIGH
      description: FeeLevel is the fee level. LOW / MEDIUM / HIGH - Defines the blockchain fee level which will be paid for the transaction.  Optional, but if provided, all other 4 fee related fields (gasLimit, gasPrice, priorityFee and maxFee) should all be null
    modelFunction:
      type: object
      properties:
        name:
          type: string
          example: approve
        stateMutability:
          type: string
          example: nonpayable
        type:
          type: string
          example: function
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/modelParamType'
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/modelParamType'
    modelParamType:
      type: object
      properties:
        components:
          type: array
          items:
            $ref: '#/components/schemas/AnyValue'
        indexed:
          type: boolean
        name:
          type: string
          example: to
        type:
          type: string
          example: address
    modelSolFile:
      type: object
      description: Source code of a contract from Etherscan.
      required:
        - fileContent
      properties:
        fileName:
          type: string
          description: Name of the file.
          example: openzeppelin-solidity/contracts/math/SafeMath.sol
        fileContent:
          type: string
          description: Content of the file. If the contract's source code was flattened before verification, this may be the entire source code.
          example: |-
            pragma solidity ^0.4.24;

            /**
             * @title SafeMath
             * @dev Math operations with safety checks...
    modelStatus:
      type: string
      enum:
        - PENDING
        - FAILED
        - COMPLETE
    modelTransactionFee:
      type: object
      properties:
        baseFee:
          type: string
        gasLimit:
          type: string
        gasPrice:
          type: string
        maxFee:
          type: string
        priorityFee:
          type: string
    modelVerificationStatus:
      type: string
      enum:
        - UNVERIFIED
        - VERIFIED
    Bytecode:
      type: string
      description: Bytecode of the contract being deployed
      minLength: 1
      writeOnly: true
      example: 0x60806040523480156200001157600080fd5b50604051806040...
    ContractName:
      type: string
      description: The contract's name.
      example: First Contract
    ContractDescription:
      type: string
      description: The description for a contract.
      example: My first hello world contract
    AbiJson:
      type: string
      description: The contract's ABI in a JSON stringified format.
      example: "[\n\t{\n\t\t\"inputs\": [],\n\t\t\"stateMutability\": \"nonpayable\",\n\t\t\"type\": \"constructor\"\n\t},\n\t..."
    WalletId:
      type: string
      format: uuid
      description: Unique identifier of the wallet that will deploy the contract.
      example: 004735f6-d9fc-44f8-933c-672cdf3d240d
    SourceAddress:
      type: string
      description: The source address of the transaction.
      example: '0x1bf9ad0cc2ad298c69a2995aa806ee832788218c'
    ConstructorParameters:
      type: array
      description: A list of arguments to pass to the contract's constructor function. Must be an empty array if there are no constructor parameters.
      items:
        $ref: '#/components/schemas/AnyValue'
      example:
        - TICK
        - 10000
    ConstructorSignature:
      type: string
      description: Signature of the constructor if the contract has one. constructor() by default.
      maxLength: 1000
      minLength: 1
      example: constructor(string ticker, uint256 totalSupply)
    Archived:
      description: Archived is the updated archived setting of the contract. If true, the contract will not be visible in your dashboard.
      type: boolean
      example: false
    AnyValue:
      description: Can be any value - string, number, boolean, array or object.
    AbiFunctionSignature:
      type: string
      description: AbiFunctionSignature defines the name and parameter types of the ABI function.
      example: safeMint(address, uint256)
    AbiParameters:
      type: array
      description: AbiParameters is the values to pass to the ABI function.
      items:
        $ref: '#/components/schemas/AnyValue'
      example:
        - '0x6E5eAf34c73D1CD0be4e24f923b97CF38e10d1f3'
        - 1
    RefId:
      type: string
      description: RefID is a custom label field.
      example: d620d523-b5ab-4541-94b0-efe4f4c825c8
...
    Error:
      type: object
      properties:
        code:
          type: integer
          description: Code is the error code
        message:
          type: string
          description: Message is the error message
...
    Blockchain:
      type: string
      enum:
        - ETH-GOERLI
        - ETH
        - AVAX-FUJI
        - AVAX
        - MATIC-MUMBAI
        - MATIC
    UpdateDate:
      type: string
      format: date-time
      description: Last update date of the resource. ISO-8601 UTC date/time.
      example: '2023-01-01T12:04:05Z'
...
    schemas_Id:
      type: string
      format: uuid
      description: Unique system generated identifier of the resource.
...
    schemas_CreateDate:
      type: string
      format: date-time
      description: Date and time the resource was created. ISO-8601 UTC date/time.
      example: '2023-01-01T12:04:05Z'
...
    schemas_IdempotencyKey:
      type: string
      description: Universally unique identifier (UUID v4) idempotency key. This key is utilized to ensure exactly-once execution of mutating requests. To create a UUIDv4 go to [uuidgenerator.net](https://www.uuidgenerator.net).
      example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
      format: uuid
    EntitySecretCiphertext:
      type: string
      description: |
        A base64 string expression of the entity secret ciphertext. The entity secret should be encrypted by the entity public key. Circle mandates that the entity secret ciphertext is unique for each API request.
      format: byte
      example: 0NtD3d3+nmgb4GqYQXzAjKF8h5Zq6sHM2k/7UQz5OIXWnU7H8O9XvVYXqDnw1wEmY...
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PREFIX:ID:SECRET
      description: Circle's API Keys are formatted in the following structure "PREFIX:ID:SECRET". All three parts are requred to make a successful request.
  headers:
    XRequestId:
      description: Universally unique identifier (UUID v4) for the request. Helpful for identifying a request when communicating with Circle support.
      schema:
        type: string
        format: uuid
        example: 2adba88e-9d63-44bc-b975-9b6ae3440dde
...